some shader and transform stuff

+++ engine/API/Vectors.h
@@ -7,8 +7,8 @@ namespace Engine {
     typedef glm::vec2 Vector2f;
     typedef glm::vec3 Vector3f;
     typedef glm::vec4 Vector4f;
-    typedef glm::vec2 Vector2i;
-    typedef glm::vec3 Vector3i;
-    typedef glm::vec4 Vector4i;
+    typedef glm::ivec2 Vector2i;
+    typedef glm::ivec3 Vector3i;
+    typedef glm::ivec4 Vector4i;
 
 }
